Move tags/metadata from domain to domain_metadata
